updated era contract

+++ contracts/ERATypes.sol
@@ -7,9 +7,9 @@ struct Marketplace {
     uint volume;
     uint64 listed;
     uint offered;
-    uint auctioned;
+    uint64 auctioned;
     address owner;
-    uint nextApplicationId;
+    uint64 nextApplicationId;
 }
 
 struct List {
@@ -18,14 +18,14 @@ struct List {
     address nftAddress;
     uint64 tokenId;
     address paymentToken;
-    uint64 ask;
+    uint128 ask;
     address owner;
     uint offers;
     bool active;
 }
 
 struct Offer {
-    uint offer_id;
+    uint64 offer_id;
     address offerer;
     address paymentToken;
     uint64 offerPrice;
@@ -33,18 +33,18 @@ struct Offer {
 }
 
 struct AuctionItem {
-    uint auctionId;
+    uint64 auctionId;
     address nftAddress;
     uint64 tokenId;
     address paymentToken;
-    uint minBid;
-    uint minBidIncrement;
-    uint startTime;
-    uint expirationTime;
+    uint32 minBid;
+    uint32 minBidIncrement;
+    uint32 startTime;
+    uint32 expirationTime;
     address owner;
     address seller;
     address highestBidder;
-    uint highestBid;
+    uint128 highestBid;
     bool active;
 }
